Rename files {parse, parser} = {parser, parse}
